+++ src/main/java/com/github/project3/controller/ImageController.java
@@ -1,6 +1,6 @@
-package com.github.project3.controller.mypage;
+package com.github.project3.controller;
 
-import com.github.project3.service.mypage.ImageService;
+import com.github.project3.service.ImageService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
